Use fixed size integers instead of usize where appropriate

+++ src/chacha.rs
@@ -16,7 +16,7 @@ use {Rng, SeedableRng, Rand, w32};
 
 const KEY_WORDS    : usize =  8; // 8 words for the 256-bit key
 const STATE_WORDS  : usize = 16;
-const CHACHA_ROUNDS: usize = 20; // Cryptographically secure from 8 upwards as of this writing
+const CHACHA_ROUNDS: u32 = 20; // Cryptographically secure from 8 upwards as of this writing
 
 /// A random number generator that uses the ChaCha20 algorithm [1].
 ///
